Add files via upload
A To-Do List is a tool to manage tasks, activities and projects. It provides a simple and effective way to keep track of everything that needs to be done, prioritize tasks, and ensure that deadlines are met.
It is a essential tool for anyone looking to stay organized, manage their time effectively, and ensure that tasks are completed on time. By providing a centralized location for all tasks and projects, it can help users stay focused, reduce stress, and increase productivity.

Features :
1. Task creation
2. Prioritization